+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- recording: session replay respect feature flag variants ([#197](https://github.com/PostHog/posthog-android/pull/197))
+
 ## 3.8.1 - 2024-10-09
 
 - recording: `OnTouchEventListener` try catch guard to swallow unexpected errors take 2 ([#196](https://github.com/PostHog/posthog-android/pull/196))